requirements.txt
Flask
sentence-transformers
faiss-cpu
numpy

